feat: add tests for email registration and login
