+++ app/src/main/java/com/epfl/drawyourpath/login/RegisterActions.kt
@@ -1,14 +1,17 @@
 package com.epfl.drawyourpath.login
 
+import androidx.fragment.app.activityViewModels
+import com.epfl.drawyourpath.R
 import android.os.Bundle
 import android.view.View
 import android.widget.Button
+import android.widget.EditText
 import androidx.fragment.app.Fragment
-import androidx.fragment.app.activityViewModels
-import com.epfl.drawyourpath.R
 
 interface RegisterActivityListener {
     fun registerWithGoogle();
+    fun registerAnonymously()
+    fun registerWithEmailAndPassword(email: String, password: String)
 }
 
 class RegisterActions : Fragment(R.layout.fragment_register_actions) {
@@ -26,12 +29,24 @@ class RegisterActions : Fragment(R.layout.fragment_register_actions) {
         // events.
         assert(activity is RegisterActivityListener);
 
+        val emailInput    = view.findViewById<EditText>(R.id.ET_Email)
+        val passwordInput = view.findViewById<EditText>(R.id.ET_Password)
+
         val registerButton = view.findViewById<Button>(R.id.BT_Login);
-        registerButton?.setOnClickListener { viewModel.showLoginUI() }
+        registerButton?.setOnClickListener {
+            viewModel.showLoginUI()
+        }
 
         val registerWithGoogleButton = view.findViewById<Button>(R.id.BT_RegisterGoogle)
         registerWithGoogleButton.setOnClickListener {
             getRegisterActivity().registerWithGoogle()
+        }
+
+        val registerWithEmailButton = view.findViewById<Button>(R.id.BT_RegisterEmail)
+        registerWithEmailButton.setOnClickListener {
+            getRegisterActivity().registerWithEmailAndPassword(
+                emailInput.text.toString(),
+                passwordInput.text.toString())
         }
     }
 }
